# src/main/java/br/com/senai/autoescolas164/adapter/in/controller/InstrucaoInController/InstrucaoController.java
package br.com.senai.autoescolas164.adapter.in.controller.InstrucaoInController;

import br.com.senai.autoescolas164.adapter.in.controller.InstrucaoInController.request.instrucao.DadosAgendamento;
import br.com.senai.autoescolas164.adapter.in.controller.InstrucaoInController.response.instrucao.DadosDetalhamentoAgendamento;
import br.com.senai.autoescolas164.application.service.AgendaDeInstrucoes;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/instrucoes")
@RequiredArgsConstructor
public class InstrucaoController {
    private final AgendaDeInstrucoes agenda;

    @PostMapping
    public ResponseEntity<DadosDetalhamentoAgendamento> agendarInstrucao(
            @RequestBody @Valid DadosAgendamento dados) {
        return ResponseEntity.ok(agenda.agendar(dados));
    }
}